scripts/requirements-discovery.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0
PyYAML>=6.0.0
pytest>=8.3.0
responses>=0.25.0
websocket-client>=1.8.0
